fix: fix toTransitionName()
While fixing mistakes in toTransiotionName(), refactor
toTransitionName() into a new file alone.
